Reduce app graph code duplication
